+++ runtime-modules/proposals/discussion/src/types.rs
@@ -1,13 +1,8 @@
-use crate::Trait;
-
 use codec::{Decode, Encode};
 #[cfg(feature = "std")]
 use serde::{Deserialize, Serialize};
 
-use rstd::marker::PhantomData;
 use rstd::prelude::*;
-
-use system::ensure_signed;
 
 /// Represents a discussion thread
 #[cfg_attr(feature = "std", derive(Serialize, Deserialize, Debug))]
@@ -72,52 +67,5 @@ impl<ThreadAuthorId: Clone> ThreadCounter<ThreadAuthorId> {
             author_id,
             counter: 1,
         }
-    }
-}
-
-/// Abstract validator for the origin(account_id) and actor_id (eg.: thread author id).
-pub trait ActorOriginValidator<Origin, ActorId> {
-    /// Check for valid combination of origin and actor_id
-    fn ensure_actor_origin(origin: Origin, actor_id: ActorId, error: &'static str) -> Result<(), &'static str>;
-}
-
-// Member of the Joystream organization
-pub(crate) type MemberId<T> = <T as membership::members::Trait>::MemberId;
-
-/// Default discussion system actor origin validator. Valid for both thread and post authors.
-pub struct MembershipOriginValidator<T> {
-    marker: PhantomData<T>,
-}
-
-impl<T> MembershipOriginValidator<T> {
-    /// Create ThreadPostActorOriginValidator instance
-    pub fn new() -> Self {
-        MembershipOriginValidator {
-            marker: PhantomData,
-        }
-    }
-}
-
-impl<T: Trait> ActorOriginValidator<<T as system::Trait>::Origin, MemberId<T>>
-    for MembershipOriginValidator<T>
-{
-    /// Check for valid combination of origin and actor_id. Actor_id should be valid member_id of
-    /// the membership module
-    fn ensure_actor_origin(origin: <T as system::Trait>::Origin, actor_id: MemberId<T>, error : &'static str) -> Result<(), &'static str> {
-        // check valid signed account_id
-        let account_id = ensure_signed(origin)?;
-
-        // check whether actor_id belongs to the registered member
-        let profile_result = <membership::members::Module<T>>::ensure_profile(actor_id);
-
-        if let Ok(profile) = profile_result {
-            // whether the account_id belongs to the actor
-            if  profile.root_account == account_id
-                || profile.controller_account == account_id {
-                return Ok(())
-            }
-        }
-
-        Err(error)
     }
 }
